core: Make pkt_size and pkt_pad size_t
Fix bad signedness comparison:

```
core.c: In function ‘mctp_binding_set_tx_enabled’:
core.c:630:25: error: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Werror=sign-compare]
  630 |   if (binding->pkt_size < MCTP_PACKET_SIZE(MCTP_BTU)) {
      |                         ^
cc1: all warnings being treated as errors
```

+++ libmctp.h
@@ -102,8 +102,8 @@ struct mctp_binding {
 	uint8_t version;
 	struct mctp_bus *bus;
 	struct mctp *mctp;
-	int pkt_size;
-	int pkt_pad;
+	size_t pkt_size;
+	size_t pkt_pad;
 	int (*start)(struct mctp_binding *binding);
 	int (*tx)(struct mctp_binding *binding, struct mctp_pktbuf *pkt);
 	mctp_rx_fn control_rx;
